Improved set_cursor handling
Move druid-shell handling from WinHandle to WinCtx.

Add a `set_cursor` method to `EventCtx`. Previously you'd get a window
handle and call `set_cursor` on that, but it's been removed.

The `EventCtx::set_cursor` method doesn't immediately delegate it
to the WinCtx, but instead defers it to the end of event propagation.

Don't set a default cursor in the window class on Windows (this
doesn't work). Instead expect it to be set in the mouse move handler,
and set the default to arrow when dispatching the mouse move event.