ci: Do less work in CI
This reduces the amount of work we will do in CI in two ways:

- it moves the 'deny warnings' check to happen in the same
job as the `cargo test` check, because they both compile
at the same optimization level and we can reuse artifacts.

- it removes the 'cargo check' job, since any failure there
will also be caught in `cargo test`. If jobs were sequential
it might be worth checking first, to fail quickly, but
since jobs are parallel this is just extra work for no
tangible benefit.