[shell] Rewrite clipboard API
This was a slog, and there are problems, but it will get runebender
moving forward.

[mac] First cut of custom clipboard paste types

This checkpoint attempts to generalize the idea of pasting custom
types on macOS in a way that I *hope* is fairly flexible.

I don't expect this to last, because I would like to have a common
mechanism for reading and writing of custom types to the clipboard,
but I'd like to check this in before I make too much of a mess.

Impl most of a new clipboard API

This is the bones of a more flexible system for copy/paste. The most
basic idea is that when the user wants to copy or paste custom types,
they define some marker struct and implement the `ClipboardRead` and or
`ClipboardWrite` traits.

Methods on these traits can be implemented per-platform, to return
platform specific information about how to serialize and deserialize
a given type. If impls are omitted for a platform, the type is
ignored on that platform.

The intent with this API is to be maxmimally accomodating to the fact
that copy/paste APIs are very different on different platforms;
instead of trying to come up with a lowest common feature set,
we instead try to be as general as possible, without sacrificing
ergonomics.

TODOs:
- I would like to prove this out on at least one additional platform
- I would like to make copy/cut easier to use with this API; in
particular I would like to support copying multiple data-types.

[mac] Support multiple format cut/copy

[win] Get basic cut/copy working with new API

[win] Preliminary copy/paste of custom formats and multi-format

[shell] Redo clipboard code

Ultimately I'm not entirely happy with this API, but eventually
you have to move on and live your life.