Implement 'save as' dialog, tweak open/save API
This gets the 'save as' dialog working in a similar fashion to
the 'open' dialog.

The API isn't great. In druid-shell, open and save_as are exposed
as two different methods on WinCtx. They could be unified, now,
at the cost of a slightly less ergonomic API; the caller would
have to bring the FileDialogType enum into scope and pass it
directly.

In druid, the API is also slightly different. The Event::Open variant
has been removed, and both open and save are now fully Command based.

The basic idea is that there are two 'special' commands,
SHOW_SAVE_PANEL and SHOW_OPEN_PANEL, that are handled by the framework.
When these commands are received, the framework handles showing the
appropriate panel, and getting the input. If this succeeds,
the framework then sends either the SAVE_FILE or OPEN_FILE command,
which is expected to be handled by the framework author.