Add Event::Zoom and WinHandler::zoom (for pinch events)
Pinch to zoom is a special case of gesture, which has (at least on
macOS) special treatment through NSEvent. At some point we will need
some sort of general mechanism for handling trackpad events,
but for now it seems reasonable to just handle this as a special event.

This only includes an implementation for macOS; I haven't looked into
the other platforms.