Move WindowHandle and event handling logic into Window struct
This has two motivations.

Moving WindowHandle into the Window struct: win_handler.rs has
included a bunch of weird bookkeeping code intended to ensure
that we always have both a window and a handle. This patch
hopefully simplifies that, by making a distinction between a
'PendingWindow', that has been created in druid but not connected to
by the platform, and a 'Window', which is _always_ backed by a handle.
This means we don't need to check that we have both a window and a
handle; once a Window exists, it means that its handle also exists.
This hopefully makes the code cleaner and easier to follow along with.

Moving base event handling logic into Window struct: this is motivated
by wanting better testing. We test at the window level, so any logic
that exists in the Window struct is exercised by our tests; logic
outside of this (such as in DruidHandler and AppState) is not.

By moving more logic into the Window, our tests better reflect the
actual behaviour of the software.