Make LifeCycle::Register be fully incremental
This moves to storing the intermediate state for widget focusing
in `BaseState`, so that it does not need to always be recomputed
from scratch anytime a child changes.

Along the way it also moves to making LifeCycleCtx take a reference
to BaseState, instead of redundantly storing its own copies of
everything.

It also fixes a bug, surfaced in testing, whereby widgets added during
an event() call would not receive WidgetAdded or Register before update().