Rework list example for clarity
The existing indexing for the list example used the length of the vector
which meant you could start with [1, 2, 3], delete [2], then your next
add would result in [1, 3, 3]. Deleting either [3] would delete both due
to the value matching.

This change makes the left and right lists maintain their own
monotonically increasing indices so the intent of '1 add' and '1 delete'
is clear.