Restore broken rustfmt
Switching to the cfg_if! macro broke rustfmt in the druid-shell platform
modules. This patch restores the individual #[cfg] attributes on the
items, and also applies rustfmt, demonstrating how much entropy can
accumulate if you let it.

This is the first of two commits. It makes the cfg_if! change but
doesn't fix the formatting.