Update to latest x11rb (#2231)
* Update to x11rb 0.9.0

The only necessary change was in error handling: X11Error now contains
more information about the request that caused the error. Namely, the
name of its X11 extension is saved in an Option<String> and this is not
Copy, so we need to .clone().

Signed-off-by: Uli Schlachter <psychon@znc.in>

* Update to x11rb 0.10

x11rb was split into two crates, x11rb and x11rb-protocol. This is
pretty transparent for us except that x11rb::resource_manager::Database
is now a re-export of x11rb_protocol::resource_manager::Database and
does not have a new_from_default() associated method. Thus, we have to
switch to x11rb::resource_manager::new_from_default() for the same
functionality.

Signed-off-by: Uli Schlachter <psychon@znc.in>