Remove outdated documentation in LifeCycle events (#2320)
The documentation previously included an outdated claim that users
interested in automatic focus should call `register_for_focus()` in
response to `LifeCycle::WidgetAdded`. This is no longer true since the
addition of `LifeCycle::BuildFocusChain`, so the relevant section was
deleted. Additionally, this commit fixes a minor spelling error in the
documentation of the latter.

Co-authored-by: Maeve Sproule <code@sprock.dev>